binary xml

+++ TinyBinaryXml/TbXmlAttributeType.cs
@@ -4,7 +4,7 @@ namespace TinyBinaryXml
 {
 	public enum TB_XML_ATTRIBUTE_TYPE
 	{
-		FLOAT = 0, 
+		DOUBLE = 0, 
 		STRING
 	}
 }
